+++ integration-tests/emulated/tests/collectives/collectives-polkadot/src/tests/fellowship_treasury.rs
@@ -17,20 +17,16 @@ use crate::*;
 use asset_hub_polkadot_runtime::xcm_config::LocationToAccountId as AssetHubLocationToAccountId;
 use emulated_integration_tests_common::accounts::ALICE;
 use frame_support::{
-	assert_ok, dispatch::RawOrigin, instances::Instance1, sp_runtime::traits::Dispatchable,
-	traits::fungible::Inspect,
+	assert_ok, instances::Instance1, sp_runtime::traits::Dispatchable, traits::fungible::Inspect,
 };
-use polkadot_runtime::OriginCaller;
+use parachains_common::pay::VersionedLocatableAccount;
 use polkadot_runtime_common::impls::VersionedLocatableAsset;
 use polkadot_runtime_constants::currency::UNITS;
-use polkadot_system_emulated_network::polkadot_emulated_chain::polkadot_runtime::Dmp;
 use xcm_executor::traits::ConvertLocation;
 
-// Fund Fellowship Treasury from Polkadot Treasury and spend from Fellowship Treasury.
+// Fund Fellowship Treasury from Asset Hub Treasury and spend from Fellowship Treasury.
 #[test]
 fn fellowship_treasury_spend() {
-	// initial treasury balance on Asset Hub in DOTs.
-	let treasury_balance = 20_000_000 * UNITS;
 	// target fellowship balance on Asset Hub in DOTs.
 	let fellowship_treasury_balance = 1_000_000 * UNITS;
 	// fellowship first spend balance in DOTs.
@@ -42,89 +38,45 @@ fn fellowship_treasury_spend() {
 		)
 	});
 
-	Polkadot::execute_with(|| {
-		type RuntimeEvent = <Polkadot as Chain>::RuntimeEvent;
-		type RuntimeCall = <Polkadot as Chain>::RuntimeCall;
-		type Runtime = <Polkadot as Chain>::Runtime;
-		type Balances = <Polkadot as PolkadotPallet>::Balances;
-		type Treasury = <Polkadot as PolkadotPallet>::Treasury;
+	AssetHubPolkadot::execute_with(|| {
+		type RuntimeEvent = <AssetHubPolkadot as Chain>::RuntimeEvent;
+		type Balances = <AssetHubPolkadot as AssetHubPolkadotPallet>::Balances;
+
+		let root = <AssetHubPolkadot as Chain>::RuntimeOrigin::root();
+		let treasury_account = asset_hub_polkadot_runtime::Treasury::account_id();
 
 		// Fund Treasury account on Asset Hub with DOTs.
-
-		let root = <Polkadot as Chain>::RuntimeOrigin::root();
-		let treasury_account = Treasury::account_id();
-
-		// Mint assets to Treasury account on Relay Chain.
 		assert_ok!(Balances::force_set_balance(
 			root.clone(),
 			treasury_account.clone().into(),
-			treasury_balance * 2,
+			fellowship_treasury_balance * 2,
 		));
-		Dmp::make_parachain_reachable(1000);
 
-		let native_asset = Location::here();
-		let asset_hub_location: Location = [Parachain(1000)].into();
-		let treasury_location: Location = (Parent, PalletInstance(19)).into();
-
-		let teleport_call = RuntimeCall::Utility(pallet_utility::Call::<Runtime>::dispatch_as {
-			as_origin: bx!(OriginCaller::system(RawOrigin::Signed(treasury_account))),
-			call: bx!(RuntimeCall::XcmPallet(pallet_xcm::Call::<Runtime>::teleport_assets {
-				dest: bx!(VersionedLocation::from(asset_hub_location.clone())),
-				beneficiary: bx!(VersionedLocation::from(treasury_location)),
-				assets: bx!(VersionedAssets::from(Assets::from(Asset {
-					id: native_asset.clone().into(),
-					fun: treasury_balance.into()
-				}))),
-				fee_asset_item: 0,
-			})),
-		});
-
-		// Dispatched from Root to `despatch_as` `Signed(treasury_account)`.
-		assert_ok!(teleport_call.dispatch(root));
-
-		assert_expected_events!(
-			Polkadot,
-			vec![
-				RuntimeEvent::XcmPallet(pallet_xcm::Event::Sent { .. }) => {},
-			]
-		);
-	});
-
-	Polkadot::execute_with(|| {
-		type RuntimeEvent = <Polkadot as Chain>::RuntimeEvent;
-		type RuntimeCall = <Polkadot as Chain>::RuntimeCall;
-		type RuntimeOrigin = <Polkadot as Chain>::RuntimeOrigin;
-		type Runtime = <Polkadot as Chain>::Runtime;
-		type Treasury = <Polkadot as PolkadotPallet>::Treasury;
-
-		// Fund Fellowship Treasury from Polkadot Treasury.
-
-		let treasury_origin: RuntimeOrigin =
-			polkadot_runtime::governance::pallet_custom_origins::Origin::Treasurer.into();
 		let fellowship_treasury_location: Location =
 			Location::new(1, [Parachain(1001), PalletInstance(65)]);
-		let asset_hub_location: Location = [Parachain(1000)].into();
-		let native_asset_on_asset_hub = Location::parent();
 
-		let treasury_spend_call = RuntimeCall::Treasury(pallet_treasury::Call::<Runtime>::spend {
-			asset_kind: bx!(VersionedLocatableAsset::V5 {
-				location: asset_hub_location.clone(),
-				asset_id: native_asset_on_asset_hub.into(),
+		// Spend from Asset Hub Treasury to Fellowship Treasury.
+		assert_ok!(asset_hub_polkadot_runtime::Treasury::spend(
+			root,
+			Box::new(VersionedLocatableAsset::V5 {
+				location: Location::new(0, []),
+				asset_id: Location::parent().into(),
 			}),
-			amount: fellowship_treasury_balance,
-			beneficiary: bx!(VersionedLocation::from(fellowship_treasury_location)),
-			valid_from: None,
-		});
-
-		assert_ok!(treasury_spend_call.dispatch(treasury_origin));
+			fellowship_treasury_balance,
+			Box::new(VersionedLocatableAccount::V5 {
+				location: Location::new(0, []),
+				account_id: fellowship_treasury_location,
+			}),
+			None,
+		));
 
 		// Claim the spend.
-
-		let alice_signed = RuntimeOrigin::signed(Polkadot::account_id_of(ALICE));
-		assert_ok!(Treasury::payout(alice_signed.clone(), 0));
+		let alice_signed =
+			<AssetHubPolkadot as Chain>::RuntimeOrigin::signed(Polkadot::account_id_of(ALICE));
+		assert_ok!(asset_hub_polkadot_runtime::Treasury::payout(alice_signed, 0));
 
 		assert_expected_events!(
-			Polkadot,
+			AssetHubPolkadot,
 			vec![
 				RuntimeEvent::Treasury(pallet_treasury::Event::AssetSpendApproved { .. }) => {},
 				RuntimeEvent::Treasury(pallet_treasury::Event::Paid { .. }) => {},
@@ -133,11 +85,9 @@ fn fellowship_treasury_spend() {
 	});
 
 	AssetHubPolkadot::execute_with(|| {
-		type RuntimeEvent = <AssetHubPolkadot as Chain>::RuntimeEvent;
 		type Balances = <AssetHubPolkadot as AssetHubPolkadotPallet>::Balances;
 
 		// Ensure that the funds deposited to the Fellowship Treasury account.
-
 		let fellowship_treasury_location: Location =
 			Location::new(1, [Parachain(1001), PalletInstance(65)]);
 		let fellowship_treasury_account =
@@ -146,19 +96,6 @@ fn fellowship_treasury_spend() {
 		assert_eq!(
 			<Balances as Inspect<_>>::balance(&fellowship_treasury_account),
 			fellowship_treasury_balance
-		);
-
-		// Assert events triggered by xcm pay program:
-		// 1. treasury asset transferred to spend beneficiary;
-		// 2. response to Relay Chain Treasury pallet instance sent back;
-		// 3. XCM program completed;
-		assert_expected_events!(
-			AssetHubPolkadot,
-			vec![
-				RuntimeEvent::Balances(pallet_balances::Event::Transfer { .. }) => {},
-				RuntimeEvent::ParachainSystem(cumulus_pallet_parachain_system::Event::UpwardMessageSent { .. }) => {},
-				RuntimeEvent::MessageQueue(pallet_message_queue::Event::Processed { success: true ,.. }) => {},
-			]
 		);
 	});
 
@@ -222,7 +159,7 @@ fn fellowship_treasury_spend() {
 
 		// Assert events triggered by xcm pay program:
 		// 1. treasury asset transferred to spend beneficiary;
-		// 2. response to Relay Chain Treasury pallet instance sent back;
+		// 2. response to Collectives chain Fellowship Treasury pallet instance sent back;
 		// 3. XCM program completed;
 		assert_expected_events!(
 			AssetHubPolkadot,